fix: sharelink swiftlint fixes
